Gross fix for gross double-revision issue
We can't rely on only
`select("core/editor").getCurrentPostLastRevisionId`. Because we have to
disable `wp_save_post_revision_check_for_changes` in order to save
revisions separately, two revision saves are triggered with every save,
and meta updates (including signatures) only happen on the second, and
`getCurrentPostLastRevisionId` only gets updated after the first. This
means the first revision may not have latest metadata and/or invalid
signatures, and this is the revision ID that will be indexed on
blockchain.

As a workaround, this creates an endpoint to fetch actual latest
revision ID from server, and adds our own store for this value. In
`effects` we subscribe to gutenberg store, and after meta box save we
make sure to fetch latest revision ID from our endpoint.