Ensure that search user name is cleared if address is changed
- Search for and select an autocompleted user that has an address
- Change the address
- The name field should be reset to empty

The last wasn't happening, this commit fixes, though why this change
makes it happen isn't totally clear and I didn't bother looking.

The reason this matters is that if it's left not blank, it gives the
impression that adding this wallet address will *update* the wallet
address for the user whose name is in the box, but it actually won't.